Revert "merge"
This reverts commit 66e27358dab97650356b4f370a0cfd71f47f6ae7, reversing
changes made to 507150b3e6307950557b1b4b73595134156dc0ba.

+++ software/src/main/java/isd/aims/main/InterbankSubsystem/vnPay/VnPaySubsystemController.java
@@ -123,7 +123,7 @@ public class VnPaySubsystemController {
         String errorCode = response.getVnp_TransactionStatus();
         String transactionId = response.getVnp_TransactionNo();
         String transactionContent = response.getVnp_OrderInfo();
-        long amount = Long.parseLong(response.getVnp_Amount()) / 100;
+        int amount = Integer.parseInt(response.getVnp_Amount()) / 100;
         String createdAt = response.getVnp_PayDate();
         SimpleDateFormat dateFormat = new SimpleDateFormat("yyyyMMddHHmmss");
